Separate context hooks into dedicated files and improve Date mocking in tests
