+++ src/context/ProjectContext.tsx
@@ -1,18 +1,11 @@
-import { createContext, useContext, useState, ReactNode } from 'react';
+import { createContext, useState, ReactNode } from 'react';
 import { Project, ProjectContextType } from './projectTypes';
 
-const ProjectContext = createContext<ProjectContextType | undefined>(undefined);
-
-export function useProject() {
-  const context = useContext(ProjectContext);
-  if (!context) {
-    throw new Error('useProject must be used within a ProjectProvider');
-  }
-  return context;
-}
+export const ProjectContext = createContext<ProjectContextType | undefined>(undefined);
 
 export function ProjectProvider({ children }: { children: ReactNode }) {
   const [project, setProject] = useState<Project | null>(null);
+
   return (
     <ProjectContext.Provider value={{ project, setProject }}>
       {children}
