+++ src/context/useProject.tsx
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { ProjectContext } from './ProjectContext';
+
+export function useProject() {
+  const context = useContext(ProjectContext);
+
+  if (!context) {
+    throw new Error('useProject must be used within a ProjectProvider');
+  }
+
+  return context;
+}
