+++ src/context/useToast.tsx
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { ToastContext } from "./ToastContext";
+
+export function useToast() {
+  const context = useContext(ToastContext);
+
+  if (context === undefined) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+
+  return context;
+}
